Move shared API key helpers out of route loader
